refactor: restructure routing and database initialization
Fixes https://github.com/marmoym/marmoym/issues/68
