+++ src/constants/HttpMethod.ts
@@ -1,8 +1,8 @@
 const HttpMethod = {
-  GET: 'GET',
-  POST: 'POST',
-  PUT: 'PUT',
-  DELETE: 'DELETE',
+  GET: 'get',
+  POST: 'post',
+  PUT: 'put',
+  DELETE: 'delete',
 };
 
 export default HttpMethod;
