+++ src/entities/BaseEntity.ts
