+++ src/entities/Origin.ts
@@ -1,7 +1,7 @@
 import {Column, Entity} from 'typeorm';
 
 import BaseEntity from '@entities/BaseEntity';
-import {DB1} from '@database/db';
+import { DB1 } from '@modules/Database';
 
 @Entity({ database: DB1 })
 export default class Origin extends BaseEntity {
