+++ src/entities/Term.ts
@@ -1,27 +1,23 @@
 import {Column, Entity, JoinColumn, ManyToOne} from 'typeorm';
 
 import BaseEntity from '@entities/BaseEntity';
-import {DB1} from '@database/db';
+import { DB1 } from '@modules/Database';
 import User from '@entities/User';
 
 @Entity({ database: DB1 })
-export default class VoteInstance extends BaseEntity {
+export default class Term extends BaseEntity {
   @Column()
-  public targetType: string;
+  public label: string;
 
   @Column()
-  public targetId: number;
+  public roman: string;
+
+  @Column()
+  public status: string;
 
   @ManyToOne((type) => User)
   @JoinColumn({
     name: 'userId',
   })
   public user: User;
-
-  @Column()
-  public action: string;
-
-  @Column()
-  public status: string;
-
 };
