+++ src/entities/VoteInstance.ts
@@ -1,23 +1,27 @@
 import {Column, Entity, JoinColumn, ManyToOne} from 'typeorm';
 
 import BaseEntity from '@entities/BaseEntity';
-import {DB1} from '@database/db';
+import { DB1 } from '@modules/Database';
 import User from '@entities/User';
 
 @Entity({ database: DB1 })
-export default class Term extends BaseEntity {
+export default class VoteInstance extends BaseEntity {
   @Column()
-  public label: string;
+  public targetType: string;
 
   @Column()
-  public roman: string;
-
-  @Column()
-  public status: string;
+  public targetId: number;
 
   @ManyToOne((type) => User)
   @JoinColumn({
     name: 'userId',
   })
   public user: User;
+
+  @Column()
+  public action: string;
+
+  @Column()
+  public status: string;
+
 };
