--- src/routes/RouterBase.ts
@@ -1,7 +1,0 @@
-import { Router } from 'express';
-
-export default class RouterBase {
-  public static routes(): {} {
-    throw new Error(`routes() is not defined', ${this.toString()}`);
-  }
-};
